refactor: split color screen into separate files
